+++ server/src/main/java/com/ticketer/dtos/OrderDto.java
@@ -3,8 +3,8 @@ package com.ticketer.dtos;
 import java.util.List;
 
 public record OrderDto(
-        List<OrderItemDto> items,
-        int subtotal,
-        int total,
-        double taxRate) {
+                List<OrderItemDto> items,
+                int subtotal,
+                int total,
+                int taxRate) {
 }
