+++ server/src/main/java/com/ticketer/dtos/SettingsDto.java
@@ -3,6 +3,6 @@ package com.ticketer.dtos;
 import java.util.Map;
 
 public record SettingsDto(
-        double tax,
-        Map<String, String> openingHours) {
+                int tax,
+                Map<String, String> openingHours) {
 }
